Fixed form validation problems
Replaced FileField with TextAreaField for iputting of the DNA sequence,
fixed problems with validation of the form.
I hope to be able to make it work with FileField as then the user will
be able to upload their DNA sequence file, without it they will have to
paste it into the textbox.

Still problems with using the data obtained from the form. Am able to
transfer the data (e.g. message) into the encryption results template,
but not do anything with the data.